(MODULES-6526) Add setter for compatibility
Prior to this commit the code was added to enable
users to specify a compatibility version to create
a task. Unfortunately, there was no testing for
verifying idempotency. When run against an existing
task the resource would fail because it can not
handle the compatibility property.  This commit
ensures that the compatibility property is settable.